+++ include/LIEF/ELF/NoteDetails/core/CoreAuxv.hpp
@@ -70,7 +70,7 @@ class LIEF_API CoreAuxv : public Note {
 
   CoreAuxv(ARCH arch, Header::CLASS cls, std::string name,
            uint32_t type, description_t description) :
-    Note(std::move(name), Note::TYPE::CORE_AUXV, type, std::move(description)),
+    Note(std::move(name), Note::TYPE::CORE_AUXV, type, std::move(description), ""),
     arch_(arch), class_(cls)
   {}